Move actor files from "xengine/message" to "xengine/actor". Change CActorWorker::EnterLoop() return type from "void" to "bool"
